refactor(backscatter): factor shared marine_backscatter engine (#191)
Extract the modality-agnostic backscatter processing core out of
marine_sidescan_mosaic into a new marine_backscatter package so the sidescan
and (forthcoming) MBES/cube layers share ONE implementation:
- ProcessedAccumulator (best-source compositor, ADR-0006 D5/D7)
- grazingQuality (mid-swath grazing-angle quality; was gradingQuality inline)
- registry writer + jsonEscape (ADR-0005 provenance)
- their gtests (+ a new quality test)

marine_sidescan_mosaic now depends on marine_backscatter; sidescan_tier2_processed
consumes the shared symbols. Sidescan projection stays put (MBES is pre-
georeferenced by cube and feeds the shared core directly). No behavior change:
both packages build, gtest all green (backscatter 4/0, sidescan 26/0), and the
processed Tier-2 still emits identical Massabesic tiles + registry.

Part of #180. Closes #191.
