+++ marine_backscatter/src/quality.cpp
@@ -1,0 +1,38 @@
+// Copyright 2026 Center for Coastal and Ocean Mapping & NOAA-UNH Joint
+// Hydrographic Center, University of New Hampshire
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL
+// THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+#include "marine_backscatter/quality.hpp"
+
+#include <algorithm>
+#include <cmath>
+
+namespace marine_backscatter
+{
+
+std::uint16_t grazingQuality(double altitude, double ground_range)
+{
+  const double grazing = std::atan2(altitude, ground_range);   // rad, 90deg at nadir.
+  const double q = std::sin(2.0 * grazing);                    // peak at 45 deg.
+  const double scaled = std::clamp(q, 0.0, 1.0) * 65535.0;
+  return static_cast<std::uint16_t>(std::max(1.0, scaled));    // floor 1: real return.
+}
+
+}  // namespace marine_backscatter
