+++ marine_sidescan_mosaic/CMakeLists.txt
@@ -20,6 +20,7 @@ find_package(geographic_msgs REQUIRED)
 find_package(geodesy REQUIRED)
 find_package(marine_autonomy REQUIRED)
 find_package(marine_tiled_raster_store REQUIRED)
+find_package(marine_backscatter REQUIRED)
 find_package(rosbag2_cpp REQUIRED)
 find_package(rosbag2_storage REQUIRED)
 find_package(geometry_msgs REQUIRED)
@@ -33,7 +34,6 @@ add_library(${PROJECT_NAME}
   src/normalizer.cpp
   src/tier1.cpp
   src/decode.cpp
-  src/processed_accumulator.cpp
 )
 target_include_directories(${PROJECT_NAME} PUBLIC
   "$<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>"
@@ -103,6 +103,7 @@ ament_target_dependencies(sidescan_tier2_processed
   geographic_msgs
   marine_autonomy
   marine_tiled_raster_store
+  marine_backscatter
 )
 install(TARGETS sidescan_tier2_processed DESTINATION lib/${PROJECT_NAME})
 
@@ -129,11 +130,6 @@ if(BUILD_TESTING)
 
   ament_add_gtest(test_tier1 test/test_tier1.cpp)
   target_link_libraries(test_tier1 ${PROJECT_NAME})
-
-  ament_add_gtest(test_processed_accumulator test/test_processed_accumulator.cpp)
-  target_link_libraries(test_processed_accumulator ${PROJECT_NAME})
-  ament_target_dependencies(test_processed_accumulator
-    geographic_msgs marine_autonomy marine_tiled_raster_store)
 endif()
 
 ament_package()
